Allow trace list packing with multiple data sample types

+++ libmseed.h
@@ -577,9 +577,9 @@ extern int64_t mstl3_unpack_recordlist (MS3TraceID *id, MS3TraceSeg *seg, void *
                                         size_t outputsize, int8_t verbose);
 extern int mstl3_convertsamples (MS3TraceSeg *seg, char type, int8_t truncate);
 extern int mstl3_resize_buffers (MS3TraceList *mstl);
-extern int mstl3_pack (MS3TraceList *mstl, void (*record_handler) (char *, int, void *),
-                       void *handlerdata, int reclen, int8_t encoding,
-                       int64_t *packedsamples, uint32_t flags, int8_t verbose, char *extra);
+extern int64_t mstl3_pack (MS3TraceList *mstl, void (*record_handler) (char *, int, void *),
+                           void *handlerdata, int reclen, int8_t encoding,
+                           int64_t *packedsamples, uint32_t flags, int8_t verbose, char *extra);
 extern void mstl3_printtracelist (MS3TraceList *mstl, ms_timeformat_t timeformat,
                                   int8_t details, int8_t gaps);
 extern void mstl3_printsynclist (MS3TraceList *mstl, char *dccid, ms_subseconds_t subseconds);
